Configure how metrics are handled on error and when metrics go missing
When an error is encountered running a query, the query's metrics will be
dropped by default. Queries can be configured to preserve the old
metrics/values instead, or preserve the metrics but zero their values.

Similarly, when a query no longer produces certain metrics, the missing
metrics will be dropped by default. Queries can be configured to preserve
the missing metric's old value, or preserve the metric but zero the value.